+++ test/New-CopyActivity.Tests.ps1
@@ -6,7 +6,7 @@ Describe New-CopyActivity {
         Import-Module $PSScriptRoot\..\src\PsDataFactory.psd1 -Force -ErrorAction Stop
     }
 
-    Context AzureDatabricks {
+    Context SqlTableSource_SqlTableSink {
 
         BeforeEach {
             $LinkedServiceReference = New-AdfLinkedServiceReference -Name MyLinkedService
@@ -20,7 +20,7 @@ Describe New-CopyActivity {
 
     }
 
-    Context AzureDatabricks {
+    Context SqlTableSource_DeltaTableSink {
 
         BeforeEach {
             $LinkedServiceReference = New-AdfLinkedServiceReference -Name MyLinkedService
@@ -31,6 +31,37 @@ Describe New-CopyActivity {
         It 'has importSetting type' {
             $CopyActivity = New-AdfCopyActivity -Name MyCopyActivity -Source $Source -SourceType AzureSqlSource -Sink $Sink -SinkType AzureDatabricksDeltaLakeSink -ErrorAction Stop
             $CopyActivity.typeProperties.sink.importSettings.type | Should -Be 'AzureDatabricksDeltaLakeImportCommand'
+        }
+
+    }
+
+    Context DeltaTableSource_DelimitedTextSink {
+
+        BeforeEach {
+            $LinkedServiceReference = New-AdfLinkedServiceReference -Name MyLinkedService
+            $Source = New-AdfAzureDatabricksDeltaLakeTable -Name MyDataset -SchemaName MySchema -TableName MyTable -LinkedServiceReference $LinkedServiceReference -ErrorAction Stop
+            $Sink = New-AdfDelimitedTextDataset `
+                -Name MyDataset `
+                -LinkedServiceReference $LinkedServiceReference `
+                -Location ([PSCustomObject]@{
+                    type = "type"
+                    folderPath = "folderPath"
+                    container = "contain"
+                }) `
+                -ColumnDelimiter "`t" `
+                -EscapeChar "`\" `
+                -FirstRowAsHeader $true `
+                -QuoteChar "`"" `
+                -ErrorAction Stop
+        }
+
+        It 'throws if SinkStagingSettings is missing' {
+            {
+                New-AdfCopyActivity -Name MyCopyActivity `
+                    -Source $Source -SourceType AzureDatabricksDeltaLakeSource `
+                    -Sink $Sink -SinkType DelimitedTextSink `
+                    -ErrorAction Stop
+            } | Should -Throw "When SourceType is 'AzureDatabricksDeltaLakeSource' and SinkType is 'DelimitedTextSink', staging via 'SinkStagingSettings' must be enabled"
         }
 
     }