+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Exception: when SourceType is 'AzureDatabricksDeltaLakeSource' and SinkType is 'DelimitedTextSink', staging via 'SinkStagingSettings' must be enabled
+
+### Changed
+
+- allowed SourceType 'AzureDatabricksDeltaLakeSink' to 'AzureDatabricksDeltaLakeSource' in 'New-CopyActivity'
+- allowed SinkType 'DelimitedTextSource' to 'DelimitedTextSink' in 'New-CopyActivity'
+
 ### Fixed
 
 - Fix non-windows path separator in `Export-Pipeline`.